third_party/blink/web_tests/external/wpt/largest-contentful-paint/web-font-styled-text-resize-swap-expected.txt

This is a testharness.js-based test.
[FAIL] LCP should be updated if the web font styled text resizes to be larger during the swap period
  assert_equals: There should be 2 LCP entries. The 1st one corresponds to the system font and the 2nd the web font. expected 2 but got 1
Harness: the test ran to completion.
